removed FoldRewriter and used Rewriter instead. Moved ConstFoldInterface into a new constants module in the opts directory.
