+++ src/irbuild/mod.rs
@@ -1,7 +1,6 @@
 //! Utilities for building and modifying IR.
 
 pub mod dialect_conversion;
-pub mod fold_rewriter;
 pub mod inserter;
 pub mod listener;
 pub mod match_rewrite;
