+++ src/opts/mod.rs
@@ -1,4 +1,5 @@
 //! Optimizations
 
+pub mod constants;
 pub mod dce;
 pub mod mem2reg;
